requirements.txt
# Python dependencies for AI Document Validator
# Core image processing and computer vision
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# OCR and text processing
pytesseract==0.3.10
easyocr==1.7.0

# Scientific computing
scipy==1.11.3
scikit-image==0.21.0

# Machine learning (optional for advanced features)
scikit-learn==1.3.0
tensorflow==2.13.0

# Image analysis and forensics
exifread==3.0.0
python-magic==0.4.27

# Utilities
argparse
logging
json5